+++ app/Http/Controllers/ProductController.php
@@ -11,14 +11,49 @@ class ProductController extends Controller
 {
     use ApiResponse;
 
-    public function index() {
-        $product = Product::all();
+    public function index(Request $request) {
+        if ($request->has('search')) {
+            return $this->search($request->query('search'));
+        }
 
-        return $this->successResponse($product);
+        if ($request->has('category_id')) {
+            return $this->filterById($request->query('category_id'));
+        }
+
+        if ($request->has('min_price') && $request->has('max_price')) {
+            return $this->filterByPrice($request->query('min_price'), $request->query('max_price'));
+        }
+
+        if ($request->has('sort_by') && $request->has('order')) {
+            return $this->sortByCategory($request->query('sort_by'), $request->query('order'));
+        }
+
+
+        $products = Product::with(['categories', 'sellers'])->get();
+
+        //Classification
+        $products->map(function ($item) {
+            $rating = $item->rating;
+
+            if ($rating >= 8.5) {
+                $item->rating_class = "Top Rated";
+            } else if ($rating >= 7.0 && $rating <= 8.4) {
+                $item->rating_class = "Recommended"; 
+            } else if ($rating < 7.0) {
+                $item->rating_class = "Regular";
+            }
+
+            return $item;
+
+        });
+        
+        return $this->successResponse($products);
     }
 
     public function store(Request $request) {
         $validated = $request->validate([
+            'seller_id' => 'required|numeric',
+            'category_id' => 'required|numeric',
             'title' => 'required|string|max:255', 
             'description' => 'required|string', 
             'price' => 'required|numeric|min:0', 
@@ -42,15 +77,21 @@ class ProductController extends Controller
         if(!$product) {
             return $this->notFoundResponse("Produk tidak ditemukan");
         } else {
-            return $this->successResponse(new ProductResource($product), "Produk Berhasil Ditemukan");
+            return $this->successResponse($product, "Produk Berhasil Ditemukan");
         }
     }
 
     public function update(Request $request, int $id) {
         $validated = $request->validate([
-            'name' => 'required|string|max:100|unique:product_categories,name',
-            'description' => 'nullable|string',
-            'icon' => 'nullable|string'
+            'title' => 'required|string|max:255', 
+            'description' => 'required|string', 
+            'price' => 'required|numeric|min:0', 
+            'rating' => 'required|numeric|min:0|max:10', 
+            'category_id' => 'required|exists:product_categories,id', 
+            'file_path' => 'required|string', 
+            'thumbnail' => 'nullable|string', 
+            'download_count' => 'required|numeric|min:0',
+            'status' => 'in:active,inactive' 
         ]);
 
         $product = Product::find($id);
@@ -58,26 +99,42 @@ class ProductController extends Controller
         if(!$product) {
             return $this->notFoundResponse("Produk tidak ditemukan");
         } else {
+            if ($request->user()->id !== $product->seller_id) {
+                return $this->unauthorizedResponse();
+            }
             $product->update([
-                'name' => $validated['name'],
+                'title' => $validated['title'],
                 'description' => $validated['description'],
-                'icon' => $validated['icon']
+                'price' => $validated['price'],
+                'rating' => $validated['rating'],
+                'category_id' => $validated['category_id'],
+                'file_path' => $validated['file_path'],
+                'thumbnail' => $validated['thumbnail'] ?? null, 
+                'download_count' => $validated['download_count'],
+                'status' => $validated['status'] ?? 'active'
             ]);
             return $this->successResponse($product, "Data berhasil di-update");
         }
     }
 
-    public function destroy(int $id) {
+    public function destroy(int $id, Request $request) {
         $product = Product::find($id);
 
         if(!$product) {
             return $this->notFoundResponse("Produk tidak ditemukan");
         } else {
+
+            if ($request->user()->id !== $product->seller_id) {
+                return $this->unauthorizedResponse();
+            }
+            
             $product->delete();
             return $this->successResponse(message: "Produk Berhasil Dihapus");
         }
     }
 
+
+    //search by judul
     public function search(string $title) {
         $products = Product::where('title', 'LIKE', '%'.$title.'%')->get(); 
         if($products->isEmpty()) {
@@ -86,16 +143,18 @@ class ProductController extends Controller
         return $this->successResponse($products, "Produk berhasil ditemukan");
     }
 
+    //filter id
     public function filterById(int $id) {
-        $products = Product::where('id', $id)->get(); 
+        $products = Product::where('category_id', $id)->get(); 
         if($products->isEmpty()) {
             return $this->notFoundResponse("Produk tidak berhasil ditemukan");
         }
         return $this->successResponse($products, "Produk berhasil ditemukan");
     }
 
-    public function filterByPrice(int $min, int $max) {
-        $products = Product::whereBetween('price', [$min, $max])->get(); 
+    //filter price
+    public function filterByPrice(int $min_price, int $max_price) {
+        $products = Product::whereBetween('price', [$min_price, $max_price])->get(); 
 
         if($products->isEmpty()) {
             return $this->notFoundResponse("Produk tidak berhasil ditemukan");
@@ -104,23 +163,17 @@ class ProductController extends Controller
     }
 
 
-    // SORT
-    public function sortByCategory(string $category) {
+    // sorting
+    public function sortByCategory(string $category, string $order) {
         $products = null;
-        if($category == "rating") {
-            $products = Product::orderBy($category, 'desc'); 
-        } else if ($category == "price") {
-            $products = Product::orderBy($category, 'asc'); 
-        } else if ($category == "download_count") {
-            $products = Product::orderBy($category, 'desc'); 
-        }
+        $products = Product::orderBy($category, $order)->get(); 
         
-        if($products->isEmpty()) {
+        if(!$products || $products->isEmpty()) {
             return $this->notFoundResponse("Produk tidak berhasil ditemukan");
         }
         return $this->successResponse($products, "Produk berhasil ditemukan");
     }
 
-    //Classification
+    
 
 }
